requirements.txt
environs==14.1.0
pyTelegramBotAPI==4.26.0
loguru==0.7.3
google-cloud-dialogflow==2.38.0
vk-api==11.9.9
protobuf==5.29.3